# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DogMap API",
    "description": "API REST para sistema de locais pet-friendly.\n\n## Sobre o Projeto\nO DogMap é uma plataforma que permite aos usuários descobrir, avaliar e favoritar locais que aceitam pets.\n\n## Funcionalidades Principais\n- **Usuários**: Cadastro e gerenciamento de usuários com informações de seus pets\n- **Locais**: Catálogo de estabelecimentos pet-friendly com diferentes categorias\n- **Avaliações**: Sistema de avaliações e comentários para os locais\n- **Favoritos**: Lista personalizada de locais favoritos por usuário\n\n## Tecnologias\n- Node.js + Express\n- Dados mockados para desenvolvimento\n- Swagger para documentação",
    "version": "1.0.0",
    "contact": {
      "name": "DogMap Team",
      "email": "contato@dogmap.com",
      "url": "https://dogmap.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3009",
      "description": "Servidor de Desenvolvimento"
    },
    {
      "url": "https://api.dogmap.com",
      "description": "Servidor de Produção"
    }
  ],
...
  "paths": {
...
    "/api/users/{id}/favorites": {
      "get": {
        "summary": "Lista favoritos de um usuário",
        "description": "Retorna todos os locais favoritados por um usuário específico",
        "tags": ["Favoritos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            },
            "example": "user_123"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de favoritos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Favorite"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Número total de favoritos",
                      "example": 8
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/users/{id}/favorites/{placeId}": {
      "post": {
        "summary": "Adiciona local aos favoritos",
        "description": "Adiciona um local específico à lista de favoritos do usuário",
        "tags": ["Favoritos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            },
            "example": "user_123"
          },
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "description": "ID do local",
            "schema": {
              "type": "string"
            },
            "example": "place_456"
          }
        ],
        "responses": {
          "201": {
            "description": "Local adicionado aos favoritos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Local adicionado aos favoritos com sucesso"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Favorite"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Remove local dos favoritos",
        "description": "Remove um local específico da lista de favoritos do usuário",
        "tags": ["Favoritos"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do usuário",
            "schema": {
              "type": "string"
            },
            "example": "user_123"
          },
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "description": "ID do local",
            "schema": {
              "type": "string"
            },
            "example": "place_456"
          }
        ],
        "responses": {
          "200": {
            "description": "Local removido dos favoritos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Local removido dos favoritos com sucesso"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "string"
                        },
                        "placeId": {
                          "type": "string"
                        },
                        "place": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "type": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Place": {
        "type": "object",
        "description": "Informações do local pet-friendly",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do local",
            "example": "place_456"
          },
          "name": {
            "type": "string",
            "description": "Nome do estabelecimento",
            "example": "Pet Shop Amigo Fiel"
          },
          "type": {
            "type": "string",
            "description": "Categoria do local",
            "example": "Pet Shop",
            "enum": ["Pet Shop", "Clínica Veterinária", "Parque", "Shopping", "Praia", "Restaurante", "Hotel"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação do local",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": ["id", "name", "type"]
      },
...
      "Favorite": {
        "type": "object",
        "description": "Local favoritado por um usuário",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID do usuário",
            "example": "user_123"
          },
          "placeId": {
            "type": "string",
            "description": "ID do local favoritado",
            "example": "place_456"
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de adição aos favoritos",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": ["userId", "placeId", "place"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Dados inválidos ou obrigatórios não fornecidos",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Dados obrigatórios não fornecidos"
                },
                "message": {
                  "type": "string",
                  "example": "Nome, email e informações do pet são obrigatórios"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Usuário não encontrado"
                },
                "message": {
                  "type": "string",
                  "example": "Usuário com ID user_123 não foi encontrado"
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                },
                "message": {
                  "type": "string",
                  "example": "Ocorreu um erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT para autenticação (implementação futura)"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}